Module Parse_argv

Module to parse a Unix-style argv into a list of strings

val parse : string -> (string list, string) Stdlib.result

parse t will split the t string into whitespace-separated substrings, taking into account quoting using the backslash character.